requirements.txt
asgiref==3.10.0
click==8.3.0
croniter==6.0.0
Django==5.2.8
django-cors-headers==4.9.0
django-redis==6.0.0
django-rq==3.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
ffmpeg-python==0.2.0
future==1.0.0
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
redis==7.0.1
rq==2.6.0
six==1.17.0
sqlparse==0.5.3
whitenoise==6.11.0
